Refactoring: Move io.apiman.gateway.api.rest.contract to io.apiman.gateway.api.rest
