+++ gateway/api/rest-impl/src/main/java/io/apiman/gateway/api/rest/impl/SystemResourceImpl.java
@@ -16,7 +16,7 @@
 
 package io.apiman.gateway.api.rest.impl;
 
-import io.apiman.gateway.api.rest.contract.ISystemResource;
+import io.apiman.gateway.api.rest.ISystemResource;
 import io.apiman.gateway.engine.beans.GatewayEndpoint;
 import io.apiman.gateway.engine.beans.SystemStatus;
 
@@ -34,7 +34,7 @@ public class SystemResourceImpl extends AbstractResourceImpl implements ISystemR
     }
 
     /**
-     * @see io.apiman.gateway.api.rest.contract.ISystemResource#getStatus()
+     * @see ISystemResource#getStatus()
      */
     @Override
     public SystemStatus getStatus() {
@@ -48,7 +48,7 @@ public class SystemResourceImpl extends AbstractResourceImpl implements ISystemR
     }
 
     /**
-     * @see io.apiman.gateway.api.rest.contract.ISystemResource#getEndpoint()
+     * @see ISystemResource#getEndpoint()
      */
     @Override
     public GatewayEndpoint getEndpoint() {
