+++ gateway/api/rest/pom.xml
@@ -44,15 +44,6 @@
       <resource>
         <directory>src/main/resources</directory>
       </resource>
-      <!-- Include some of the source for this project so that it can be used by GWT projects. -->
-      <resource>
-        <directory>src/main/java</directory>
-        <includes>
-          <include>**/contract/**</include>
-          <include>**/exceptions/**</include>
-          <include>**/*.gwt.xml</include>
-        </includes>
-      </resource>
     </resources>
     <plugins>
       <plugin>
