+++ gateway/api/rest/src/main/java/io/apiman/gateway/api/rest/exceptions/GatewayApiErrorBean.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.apiman.gateway.api.rest.contract.exceptions;
+package io.apiman.gateway.api.rest.exceptions;
 
 import java.io.Serializable;
